requirements.txt
# Core dependencies
python-dotenv==1.0.0
pydantic==2.5.0

# AI/LLM Integration
openai==1.6.1
anthropic==0.8.1

# Date parsing
python-dateutil==2.8.2

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Code quality
black==23.12.1
flake8==7.0.0
